refactor(api): split playerData, notifications, accessTokens, agentKeys, access, and review slices out of store.ts
Phase 2 wave 4: move implementation code (not just types) for six slices into
per-slice files under store/slices/, following the delegation pattern
established for telemetry and oauth. InMemoryStore/FirestoreStore now hold one
field per slice and forward every interface method.

playerData further splits into player-data.ts (saves/drafts/play-affinity) and
world-entries.ts (world entries), and review further splits into review.ts
(assessments/re-review requests) and review-sweeps.ts (sweeps/scorecards), to
stay under the 500-line new-file cap. Maps that deleteAccountIdentity or
PublicationStore.listGameSlugs need to reach across slice boundaries are kept
non-private with a documented-exception comment, matching the existing
pattern.

Also fixes firestore-indexes.test.ts's collection-group query scanner to walk
subdirectories, restoring coverage of the worldEntries query that moved out
of store.ts, and updates creator-agent-key-routes.test.ts's internal-state
reach to follow creatorAgentKeys into its new home.
