+++ eslint-rules/comment-prose-baseline.json
@@ -291,8 +291,15 @@
     "apps/api/src/store/records/submission.ts": 1562,
     "apps/api/src/store/records/telemetry.test.ts": 92,
     "apps/api/src/store/records/telemetry.ts": 1136,
+    "apps/api/src/store/slices/access-tokens.ts": 120,
+    "apps/api/src/store/slices/access.ts": 242,
+    "apps/api/src/store/slices/agent-keys.ts": 282,
+    "apps/api/src/store/slices/notifications.ts": 116,
     "apps/api/src/store/slices/oauth.ts": 83,
+    "apps/api/src/store/slices/player-data.ts": 916,
+    "apps/api/src/store/slices/review-sweeps.ts": 437,
     "apps/api/src/store/slices/telemetry.ts": 167,
+    "apps/api/src/store/slices/world-entries.ts": 434,
     "apps/api/src/submission-status.test.ts": 35,
     "apps/api/src/submission-status.ts": 257,
     "apps/api/src/submissions.test.ts": 4563,
