+++ eslint-rules/module-size-baseline.json
@@ -112,7 +112,7 @@
     "apps/api/src/copilot-mcp-deploy.test.ts": 40,
     "apps/api/src/creation-limits.test.ts": 364,
     "apps/api/src/creation-limits.ts": 508,
-    "apps/api/src/creator-agent-key-routes.test.ts": 936,
+    "apps/api/src/creator-agent-key-routes.test.ts": 940,
     "apps/api/src/creator-agent-key-routes.ts": 190,
     "apps/api/src/creator-code.test.ts": 1398,
     "apps/api/src/creator-code.ts": 1316,
@@ -156,7 +156,7 @@
     "apps/api/src/example-files.ts": 198,
     "apps/api/src/feedback-themes.test.ts": 184,
     "apps/api/src/feedback-themes.ts": 227,
-    "apps/api/src/firestore-indexes.test.ts": 152,
+    "apps/api/src/firestore-indexes.test.ts": 158,
     "apps/api/src/game-follow-notify.ts": 98,
     "apps/api/src/game-follow-routes.test.ts": 233,
     "apps/api/src/game-follow-routes.ts": 87,
